deleted Frontend folder and made changed directly into frontend
