--- Frontend/layouts/default.vue
@@ -1,15 +1,0 @@
-<template>
-    <div>
-      <Navbar/>
-      <slot />
-    </div>
-  </template>
-
-<script>
-import Navbar from '@/components/NavNew.vue';
-export default {
-  components: {
-    Navbar
-  },
-}
-</script>
