--- Frontend/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "nuxt-app",
-  "private": true,
-  "type": "module",
-  "scripts": {
-    "build": "nuxt build",
-    "dev": "nuxt dev",
-    "generate": "nuxt generate",
-    "preview": "nuxt preview",
-    "postinstall": "nuxt prepare"
-  },
-  "dependencies": {
-    "@pinia-plugin-persistedstate/nuxt": "^1.2.0",
-    "@pinia/nuxt": "^0.5.1",
-    "nuxt": "^3.11.1",
-    "nuxt-material-design-icons-iconfont": "^1.0.2",
-    "pinia": "^2.1.7",
-    "vue": "^3.4.21",
-    "vue-router": "^4.3.0",
-    "vue3-lottie": "^3.3.0",
-    "vue3-toastify": "^0.2.1"
-  },
-  "devDependencies": {
-    "@nuxtjs/tailwindcss": "^6.11.4"
-  }
-}
