--- Frontend/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  // https://nuxt.com/docs/guide/concepts/typescript
-  "extends": "./.nuxt/tsconfig.json"
-}
-
-
-
-
-
-
-
