--- frontend/components/cardComp.vue
@@ -1,23 +1,0 @@
-<template>
-  <div class="flex-col items-center flex space-around">
-    <div
-      class=" hover:border-2 hover:border-slate-300 mx-[3.8rem] mt-[10rem] w-[17rem] h-[17rem] md:w-[25rem] md:h-[30rem]  border-2 rounded-sm justify-center">
-    </div>
-    <div
-      class=" main w-[17rem] md:w-[25rem] py-[1rem] md:py-[2rem] hover:border-2 hover:border-slate-300 border-black-800">
-      <div class="flex-col"></div>
-      <div class="text-red-700">T-Shirt</div>
-      <div class="text-black-700">TPC Cotton Printed Blue Shirt</div>
-      <div class="text-gray-500">TPC</div>
-      <div class="text-black-700">MRP: Rs. 399.00</div>
-    </div>
-  </div>
-</template>
-
-<script>
-export default {
-  name: 'cardComp'
-}
-</script>
-
-<style scoped></style>
