+++ frontend/layouts/default.vue
@@ -1,13 +1,11 @@
 <template>
-  <div>
-    <Navbar />
-    <AppHeader />
-    <Nuxt />
-  </div>
-</template>
+    <div>
+      <Navbar/>
+      <slot />
+    </div>
+  </template>
 
 <script>
-// import Navbar from '@/components/Nav.vue';
 import Navbar from '@/components/NavNew.vue';
 export default {
   components: {
