--- frontend/middleware/auth.js
@@ -1,5 +1,0 @@
-export default function ({ store, route, redirect }) {
-  
- 
-  
-}
