+++ frontend/pages/events/[id].vue
