+++ frontend/plugins/Vue3Lottie.client.ts
