--- frontend/services/api.js
@@ -1,10 +1,0 @@
-import axios from 'axios';
-
-const apiClient = axios.create({
-  baseURL: 'https://backend-getit.onrender.com/api/',
-  headers: {
-    'Content-Type': 'application/json',
-  },
-});
-
-export default apiClient;
